Feat/pixart ctrlnet public (#74)
Update Pixart-delta Controlnet inference.

+++ diffusion/model/nets/PixArt_blocks.py
@@ -267,7 +267,7 @@ class TimestepEmbedder(nn.Module):
         return embedding
 
     def forward(self, t):
-        t_freq = self.timestep_embedding(t, self.frequency_embedding_size)
+        t_freq = self.timestep_embedding(t, self.frequency_embedding_size).to(t.dtype)
         t_emb = self.mlp(t_freq)
         return t_emb
 
@@ -296,7 +296,7 @@ class SizeEmbedder(TimestepEmbedder):
             assert s.shape[0] == bs
         b, dims = s.shape[0], s.shape[1]
         s = rearrange(s, "b d -> (b d)")
-        s_freq = self.timestep_embedding(s, self.frequency_embedding_size)
+        s_freq = self.timestep_embedding(s, self.frequency_embedding_size).to(s.dtype)
         s_emb = self.mlp(s_freq)
         s_emb = rearrange(s_emb, "(b d) d2 -> b (d d2)", b=b, d=dims, d2=self.outdim)
         return s_emb
